frontend: refactor horse routes
