+++ apps/frontend/src/components/events/Choroby.tsx
@@ -1,4 +1,4 @@
-import HorseEventList from "../HorseEventList";
+import HorseEventList from "../konie/ListaZdarzenKonia";
 function Choroby() {
   return <HorseEventList type="choroby" />;
 }
