+++ apps/frontend/src/components/events/Leczenia.tsx
@@ -1,4 +1,4 @@
-import HorseEventList from "../HorseEventList";
+import HorseEventList from "../konie/ListaZdarzenKonia";
 function Leczenia() {
   return <HorseEventList type="leczenia" />;
 }
