+++ apps/frontend/src/components/events/Podkucia.tsx
@@ -1,4 +1,4 @@
-import HorseEventList from "../HorseEventList";
+import HorseEventList from "../konie/ListaZdarzenKonia";
 function Podkucia() {
   return <HorseEventList type="podkucia" />;
 }
