+++ apps/frontend/src/components/events/Profilaktyczne.tsx
@@ -1,4 +1,4 @@
-import HorseEventList from "../HorseEventList";
+import HorseEventList from "../konie/ListaZdarzenKonia";
 function Profilaktyczne() {
   return <HorseEventList type="zdarzenia_profilaktyczne" />;
 }
