+++ apps/frontend/src/components/events/Rozrody.tsx
@@ -1,4 +1,4 @@
-import HorseEventList from "../HorseEventList";
+import HorseEventList from "../konie/ListaZdarzenKonia";
 function Rozrody() {
   return <HorseEventList type="rozrody" />;
 }
