+++ apps/frontend/src/components/konie/ListaZdarzenKonia.tsx
