+++ apps/frontend/src/components/konie/SzczegolyKonia.tsx
