+++ apps/frontend/src/routes.ts
@@ -2,6 +2,7 @@ import {
   type RouteConfig,
   index,
   layout,
+  prefix,
   route,
 } from "@react-router/dev/routes";
 
@@ -10,13 +11,17 @@ export default [
   route("/login", "./components/Login.tsx"),
   route("/register", "./components/Register.tsx"),
   layout("./components/Layout.tsx", [
-    route("/konie", "./components/Konie.tsx"),
-    route("/konie/add", "./components/AddKonia.tsx"),
     route("/restart", "./components/Restart.tsx"),
-    route("konie/:id", "./components/KonieDetails.tsx"),
+
+    ...prefix("konie", [
+      index("./components/konie/Konie.tsx"),
+      route("add", "./components/konie/DodajKonia.tsx"),
+      route(":id/edit", "./components/konie/EdytujKonia.tsx"),
+      route(":id", "./components/konie/SzczegolyKonia.tsx"),
+    ]),
+
     route("wydarzenia", "./components/Wydarzenia.tsx"),
     route("/wydarzenia/add/:type", "./components/AddWydarzenie.tsx"),
-    route("/konie/:id/edit", "./components/EditKonia.tsx"),
     route("/wydarzenia/:id/rozrody", "./components/events/Rozrody.tsx"),
     route("/wydarzenia/:id/choroby", "./components/events/Choroby.tsx"),
     route("/wydarzenia/:id/leczenia", "./components/events/Leczenia.tsx"),
